Adjust to 'VideoController' functionality
