+++ src/main/java/me/universi/capacity/service/VideoServiceInterface.java
@@ -1,7 +1,8 @@
-package me.universi.capacitacao.service;
+package me.universi.capacity.service;
 
 import java.util.List;
-import me.universi.capacitacao.entidades.Video;
+
+import me.universi.capacity.entidades.Video;
 
 public interface VideoServiceInterface {
     List<Video> getVideosByCategory(String category);
